Implement JavaScript interface to main classes
Mostly reimplemented the functionality of the old script interface
with QJSEngine. Due to the lack of features in public Qt API, the
implementation is based on direct access to the underlying v8 API.

This is the first working version that makes it possible to create and
run PiiEngines in QML applications. Most of the features have not been
tested yet.

This commit also includes some refactoring to PiiResourceDatabase.
This was needed to make it possible to statically register
QMetaObjects to the resource database.